# docs/openapi.yaml
openapi: 3.0.3
info:
  title: パートナー協業プラットフォーム API
  description: |
    パートナー企業との協業を管理するためのREST API仕様です。

    ## 認証
    このAPIはBearer Token (JWT)認証を使用します。
    ログインAPIで取得したアクセストークンを`Authorization`ヘッダーに設定してください。

    ```
    Authorization: Bearer {accessToken}
    ```

    ## ページネーション
    一覧取得APIでは`limit`と`offset`パラメータでページネーションを制御できます。

    ## エラーレスポンス
    全てのエラーは統一フォーマットで返却されます。
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@example.com
servers:
  - url: https://api.example.com/api
    description: Production server
  - url: https://staging-api.example.com/api
    description: Staging server
  - url: http://localhost:3000/api
    description: Development server
...
paths:
...
  /projects:
    get:
      tags:
        - Projects
      summary: 案件一覧取得
      description: 案件一覧を取得します。フィルタリング、ページネーションに対応しています。
      operationId: listProjects
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
        - name: status
          in: query
          description: ステータスでフィルタ
          schema:
            $ref: '#/components/schemas/ProjectStatus'
        - name: partnerId
          in: query
          description: パートナーIDでフィルタ
          schema:
            type: string
        - name: managerId
          in: query
          description: 担当マネージャーIDでフィルタ
          schema:
            type: string
        - name: search
          in: query
          description: 案件名・説明での検索
          schema:
            type: string
        - name: startDateFrom
          in: query
          description: 開始日（以降）でフィルタ
          schema:
            type: string
            format: date
        - name: startDateTo
          in: query
          description: 開始日（以前）でフィルタ
          schema:
            type: string
            format: date
        - name: sortBy
          in: query
          description: ソート項目
          schema:
            type: string
            enum: [createdAt, updatedAt, startDate, name]
            default: createdAt
        - name: sortOrder
          in: query
          description: ソート順
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Projects
      summary: 案件作成
      description: 新規案件を作成します。
      operationId: createProject
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreateRequest'
      responses:
        '201':
          description: 作成成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /projects/{id}:
    get:
      tags:
        - Projects
      summary: 案件詳細取得
      description: 指定した案件の詳細情報を取得します。
      operationId: getProject
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectIdParam'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ProjectDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Projects
      summary: 案件更新
      description: 既存の案件を更新します。
      operationId: updateProject
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdateRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      tags:
        - Projects
      summary: 案件削除
      description: 案件を削除します（論理削除）。
      operationId: deleteProject
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectIdParam'
      responses:
        '204':
          description: 削除成功
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: 競合エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: HAS_ACTIVE_TASKS
                  message: 進行中のタスクがあるため削除できません

  # ============================================================
  # Task APIs
  # ============================================================
  /projects/{projectId}/tasks:
    get:
      tags:
        - Tasks
      summary: タスク一覧取得
      description: 指定した案件のタスク一覧を取得します。
      operationId: listTasks
      security:
        - BearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          description: 案件ID
          schema:
            type: string
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
        - name: status
          in: query
          description: ステータスでフィルタ
          schema:
            $ref: '#/components/schemas/TaskStatus'
        - name: assigneeId
          in: query
          description: 担当者IDでフィルタ
          schema:
            type: string
        - name: priority
          in: query
          description: 優先度でフィルタ
          schema:
            $ref: '#/components/schemas/Priority'
        - name: dueDateFrom
          in: query
          description: 期限日（以降）でフィルタ
          schema:
            type: string
            format: date
        - name: dueDateTo
          in: query
          description: 期限日（以前）でフィルタ
          schema:
            type: string
            format: date
        - name: isOverdue
          in: query
          description: 期限超過タスクのみ取得
          schema:
            type: boolean
        - name: sortBy
          in: query
          description: ソート項目
          schema:
            type: string
            enum: [createdAt, dueDate, priority]
            default: createdAt
        - name: sortOrder
          in: query
          description: ソート順
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags:
        - Tasks
      summary: タスク作成
      description: 新規タスクを作成します。
      operationId: createTask
      security:
        - BearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          description: 案件ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateRequest'
      responses:
        '201':
          description: 作成成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWTアクセストークン
  parameters:
    LimitParam:
      name: limit
      in: query
      description: 取得件数
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    OffsetParam:
      name: offset
      in: query
      description: 取得開始位置
      schema:
        type: integer
        minimum: 0
        default: 0
    ProjectIdParam:
      name: id
      in: path
      required: true
      description: 案件ID
      schema:
        type: string
...
  responses:
    Unauthorized:
      description: 認証エラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNAUTHORIZED
              message: 認証が必要です
...
    NotFound:
      description: リソース未検出
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: NOT_FOUND
              message: 指定されたリソースが見つかりません
    ValidationError:
      description: バリデーションエラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: VALIDATION_ERROR
              message: 入力内容に誤りがあります
              details:
                - field: email
                  message: 有効なメールアドレスを入力してください
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: エラーコード
            message:
              type: string
              description: エラーメッセージ
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                    description: フィールド名
                  message:
                    type: string
                    description: 詳細メッセージ
    Pagination:
      type: object
      properties:
        total:
          type: integer
          description: 総件数
        limit:
          type: integer
          description: 取得件数
        offset:
          type: integer
          description: 取得開始位置
        hasMore:
          type: boolean
          description: 追加データの有無
    Priority:
      type: string
      enum: [low, medium, high]
      description: 優先度

    # ============================================================
    # Authentication Schemas
    # ============================================================
...
    UserSummary:
      type: object
      properties:
        id:
          type: string
          description: ユーザーID
        email:
          type: string
          format: email
          description: メールアドレス
        name:
          type: string
          description: 氏名
        role:
          type: string
          enum: [admin, manager, member, partner]
          description: 役割

    # ============================================================
    # Project Schemas
    # ============================================================
    ProjectStatus:
      type: string
      enum: [draft, active, completed, archived]
      description: 案件ステータス
...
    Project:
      type: object
      properties:
        id:
          type: string
          description: 案件ID
        name:
          type: string
          description: 案件名
        description:
          type: string
          description: 説明
        status:
          $ref: '#/components/schemas/ProjectStatus'
        priority:
          $ref: '#/components/schemas/Priority'
        startDate:
          type: string
          format: date
          description: 開始日
        endDate:
          type: string
          format: date
          description: 終了日
        budget:
          type: integer
          description: 予算（円）
        actualCost:
          type: integer
          description: 実績費用（円）
        partner:
          $ref: '#/components/schemas/PartnerSummary'
        manager:
          $ref: '#/components/schemas/UserSummary'
        progress:
          type: integer
          minimum: 0
          maximum: 100
          description: 進捗率
        taskSummary:
          $ref: '#/components/schemas/TaskSummary'
        tags:
          type: array
          items:
            type: string
          description: タグ
        createdAt:
          type: string
          format: date-time
          description: 作成日時
        updatedAt:
          type: string
          format: date-time
          description: 更新日時
    ProjectDetail:
      allOf:
        - $ref: '#/components/schemas/Project'
        - type: object
          properties:
            members:
              type: array
              items:
                $ref: '#/components/schemas/UserSummary'
              description: メンバー一覧
            milestones:
              type: array
              items:
                $ref: '#/components/schemas/Milestone'
              description: マイルストーン一覧
            attachments:
              type: array
              items:
                $ref: '#/components/schemas/Attachment'
              description: 添付ファイル一覧
    ProjectListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Project'
        pagination:
          $ref: '#/components/schemas/Pagination'
    ProjectCreateRequest:
      type: object
      required:
        - name
        - startDate
        - endDate
        - partnerId
        - managerId
      properties:
        name:
          type: string
          maxLength: 100
          description: 案件名
        description:
          type: string
          maxLength: 2000
          description: 説明
        status:
          type: string
          enum: [draft, active]
          default: draft
          description: ステータス
        priority:
          $ref: '#/components/schemas/Priority'
        startDate:
          type: string
          format: date
          description: 開始日
        endDate:
          type: string
          format: date
          description: 終了日
        budget:
          type: integer
          description: 予算（円）
        partnerId:
          type: string
          description: パートナーID
        managerId:
          type: string
          description: マネージャーID
        memberIds:
          type: array
          items:
            type: string
          description: メンバーIDリスト
        tags:
          type: array
          items:
            type: string
          description: タグ
    ProjectUpdateRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
          description: 案件名
        description:
          type: string
          maxLength: 2000
          description: 説明
        status:
          $ref: '#/components/schemas/ProjectStatus'
        priority:
          $ref: '#/components/schemas/Priority'
        startDate:
          type: string
          format: date
          description: 開始日
        endDate:
          type: string
          format: date
          description: 終了日
        budget:
          type: integer
          description: 予算（円）
        partnerId:
          type: string
          description: パートナーID
        managerId:
          type: string
          description: マネージャーID
        memberIds:
          type: array
          items:
            type: string
          description: メンバーIDリスト
        tags:
          type: array
          items:
            type: string
          description: タグ
    TaskSummary:
      type: object
      properties:
        total:
          type: integer
          description: 総タスク数
        completed:
          type: integer
          description: 完了タスク数
        inProgress:
          type: integer
          description: 進行中タスク数
        pending:
          type: integer
          description: 保留タスク数
        overdue:
          type: integer
          description: 期限超過タスク数
    Milestone:
      type: object
      properties:
        id:
          type: string
          description: マイルストーンID
        name:
          type: string
          description: マイルストーン名
        dueDate:
          type: string
          format: date
          description: 期限日
        status:
          type: string
          enum: [pending, in_progress, completed]
          description: ステータス
    Attachment:
      type: object
      properties:
        id:
          type: string
          description: 添付ファイルID
        name:
          type: string
          description: ファイル名
        size:
          type: integer
          description: ファイルサイズ（バイト）
        mimeType:
          type: string
          description: MIMEタイプ
        uploadedAt:
          type: string
          format: date-time
          description: アップロード日時

    # ============================================================
    # Task Schemas
    # ============================================================
    TaskStatus:
      type: string
      enum: [pending, in_progress, review, completed]
      description: タスクステータス
    Task:
      type: object
      properties:
        id:
          type: string
          description: タスクID
        projectId:
          type: string
          description: 案件ID
        title:
          type: string
          description: タスク名
        description:
          type: string
          description: 説明
        status:
          $ref: '#/components/schemas/TaskStatus'
        priority:
          $ref: '#/components/schemas/Priority'
        dueDate:
          type: string
          format: date
          description: 期限日
        estimatedHours:
          type: integer
          description: 見積工数（時間）
        actualHours:
          type: integer
          description: 実績工数（時間）
        assignee:
          $ref: '#/components/schemas/UserSummary'
        tags:
          type: array
          items:
            type: string
          description: タグ
        isOverdue:
          type: boolean
          description: 期限超過フラグ
        completedAt:
          type: string
          format: date-time
          nullable: true
          description: 完了日時
        createdAt:
          type: string
          format: date-time
          description: 作成日時
        updatedAt:
          type: string
          format: date-time
          description: 更新日時
    TaskListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        pagination:
          $ref: '#/components/schemas/Pagination'
    TaskCreateRequest:
      type: object
      required:
        - title
        - dueDate
      properties:
        title:
          type: string
          maxLength: 200
          description: タスク名
        description:
          type: string
          maxLength: 4000
          description: 説明
        status:
          type: string
          enum: [pending, in_progress]
          default: pending
          description: ステータス
        priority:
          $ref: '#/components/schemas/Priority'
        dueDate:
          type: string
          format: date
          description: 期限日
        estimatedHours:
          type: integer
          description: 見積工数（時間）
        assigneeId:
          type: string
          description: 担当者ID
        tags:
          type: array
          items:
            type: string
          description: タグ
...
    PartnerSummary:
      type: object
      properties:
        id:
          type: string
          description: パートナーID
        name:
          type: string
          description: 会社名
...